mappings: change email_adresses in authors to keyword * requires remapping and reindexing of the authors
