+++ backend/inspirehep/search/mappings/v7/records/authors.json
@@ -176,7 +176,7 @@
             "type": "boolean"
           },
           "value": {
-            "type": "text"
+            "type": "keyword"
           }
         }
       },
