+++ backend/tests/integration-async/disambiguation/test_disambiguation_tasks.py
@@ -8,14 +8,15 @@
 import json
 
 from helpers.providers.faker import faker
-from helpers.utils import es_search, retry_until_matched
+from helpers.utils import es_search, retry_until_matched, retry_until_pass
 from invenio_db import db
 from invenio_pidstore.models import PersistentIdentifier
 from invenio_search import current_search
 
 from inspirehep.disambiguation.tasks import disambiguate_signatures
-from inspirehep.records.api import AuthorsRecord
+from inspirehep.records.api import AuthorsRecord, InspireRecord
 from inspirehep.records.api.literature import LiteratureRecord
+from inspirehep.search.api import InspireSearch
 
 
 def test_signature_linked_by_disambiguation_has_correct_facet_author_name(
@@ -67,3 +68,339 @@ def test_signature_linked_by_disambiguation_has_correct_facet_author_name(
         },
     ]
     retry_until_matched(steps)
+
+
+def test_disambiguation_runs_after_record_creation(
+    inspire_app, celery_app_with_context, celery_session_worker, enable_disambiguation
+):
+    author_data = faker.record("aut", with_control_number=True)
+    author_data.update(
+        {
+            "name": {"value": "Brian Gross"},
+            "ids": [{"schema": "INSPIRE BAI", "value": "J.M.Maldacena.1"}],
+            "email_addresses": [{"current": True, "value": "test@test.com"}],
+        }
+    )
+    author_record = InspireRecord.create(author_data)
+    db.session.commit()
+
+    def assert_authors_records_exist_in_es():
+        author_record_from_es = InspireSearch.get_record_data_from_es(author_record)
+        assert author_record_from_es
+
+    retry_until_pass(assert_authors_records_exist_in_es)
+
+    literature_data = faker.record("lit", with_control_number=True)
+    literature_data.update(
+        {
+            "authors": [
+                {
+                    "full_name": "Brian Gross",
+                    "ids": [{"schema": "INSPIRE BAI", "value": "J.M.Maldacena.1"},],
+                    "emails": ["test@test.com"],
+                },
+            ]
+        }
+    )
+    literature_record = LiteratureRecord.create(literature_data)
+    db.session.commit()
+
+    def assert_disambiguation_task():
+        literature_record_from_es = InspireSearch.get_record_data_from_es(
+            literature_record
+        )
+        assert (
+            str(author_data["control_number"])
+            in literature_record_from_es["authors"][0]["record"]["$ref"]
+        )
+
+    retry_until_pass(assert_disambiguation_task, retry_interval=2)
+
+
+def test_disambiguate_many_authors_runs_after_record_creation(
+    inspire_app, celery_app_with_context, celery_session_worker, enable_disambiguation
+):
+    author_1 = faker.record("aut", with_control_number=True)
+    author_1.update(
+        {
+            "name": {"value": "Brian Gross"},
+            "ids": [
+                {"schema": "INSPIRE ID", "value": "INSPIRE-00304313"},
+                {"schema": "INSPIRE BAI", "value": "J.M.Maldacena.1"},
+            ],
+            "email_addresses": [{"current": True, "value": "test@test.com"}],
+        }
+    )
+    author_2 = faker.record("aut", with_control_number=True)
+    author_2.update(
+        {
+            "name": {"value": "Donald Matthews"},
+            "ids": [{"schema": "INSPIRE BAI", "value": "H.Khalfoun.1"}],
+            "email_addresses": [
+                {"current": True, "value": "test1@test.pl"},
+                {"current": True, "value": "test1.1@test.pl"},
+            ],
+        }
+    )
+
+    author_record_1 = InspireRecord.create(author_1)
+    author_record_2 = InspireRecord.create(author_2)
+    db.session.commit()
+
+    def assert_authors_records_exist_in_es():
+        author_record_1_from_es = InspireSearch.get_record_data_from_es(author_record_1)
+        author_record_2_from_es = InspireSearch.get_record_data_from_es(author_record_2)
+        assert author_record_1_from_es and author_record_2_from_es
+
+    retry_until_pass(assert_authors_records_exist_in_es)
+
+    literature_data = faker.record("lit", with_control_number=True)
+    literature_data.update(
+        {
+            "authors": [
+                {
+                    "full_name": "Brian Gross",
+                    "ids": [
+                        {"schema": "INSPIRE ID", "value": "INSPIRE-00304313"},
+                        {"schema": "INSPIRE BAI", "value": "J.M.Maldacena.1"},
+                    ],
+                    "emails": ["test@test.com"],
+                },
+                {
+                    "full_name": "Donald Matthews",
+                    "ids": [{"schema": "INSPIRE BAI", "value": "H.Khalfoun.1"}],
+                    "emails": ["test1@test.pl", "test1.1@test.pl"],
+                },
+            ]
+        }
+    )
+    literature_record = LiteratureRecord.create(literature_data)
+    db.session.commit()
+
+    def assert_disambiguation_task():
+        literature_record_from_es = InspireSearch.get_record_data_from_es(
+            literature_record
+        )
+        literature_record_from_es_authors = literature_record_from_es.get("authors")
+        assert (
+            str(author_1["control_number"])
+            in literature_record_from_es_authors[0]["record"]["$ref"]
+        )
+        assert (
+            str(author_2["control_number"])
+            in literature_record_from_es_authors[1]["record"]["$ref"]
+        )
+
+    retry_until_pass(assert_disambiguation_task, retry_interval=2)
+
+
+def test_disambiguate_authors_doesnt_match_when_author_is_ambiguous(
+    inspire_app, enable_disambiguation
+):
+    author_1 = faker.record("aut", with_control_number=True)
+    author_1.update(
+        {
+            "name": {"value": "Brian Gross"},
+            "ids": [
+                {"schema": "INSPIRE ID", "value": "INSPIRE-00304313"},
+                {"schema": "INSPIRE BAI", "value": "J.M.Maldacena.2"},
+            ],
+            "email_addresses": [{"current": True, "value": "test@test.com"}],
+        }
+    )
+    author_2 = faker.record("aut", with_control_number=True)
+    author_2.update(
+        {
+            "name": {"value": "Brian Gross"},
+            "ids": [
+                {"schema": "INSPIRE ID", "value": "INSPIRE-00300003"},
+                {"schema": "INSPIRE BAI", "value": "J.M.Maldacena.1"},
+            ],
+            "email_addresses": [{"current": True, "value": "test@test.com"}],
+        }
+    )
+
+    author_record_1 = InspireRecord.create(author_1)
+    author_record_2 = InspireRecord.create(author_2)
+    db.session.commit()
+
+    authors = [{"full_name": "Brian Gross", "emails": ["test@test.com"],}]
+
+    literature_data = faker.record("lit", with_control_number=True)
+    literature_data.update({"authors": authors})
+    literature_record = LiteratureRecord.create(literature_data)
+    db.session.commit()
+
+    def assert_disambiguation_task():
+        literature_record_from_es = InspireSearch.get_record_data_from_es(
+            literature_record
+        )
+        assert not literature_record_from_es["authors"][0].get("record")
+
+    retry_until_pass(assert_disambiguation_task)
+
+
+def test_disambiguation_doesnt_run_with_feature_flag_disabling_it(
+    inspire_app, celery_app_with_context, celery_session_worker
+):
+    author_data = faker.record("aut", with_control_number=True)
+    author_data.update(
+        {
+            "name": {"value": "Brian Gross"},
+            "ids": [{"schema": "INSPIRE BAI", "value": "J.M.Maldacena.1"},],
+            "email_addresses": [{"current": True, "value": "test@test.com"}],
+        }
+    )
+    author_record = InspireRecord.create(author_data)
+    db.session.commit()
+
+    def assert_authors_records_exist_in_es():
+        author_record_from_es = InspireSearch.get_record_data_from_es(author_record)
+        assert author_record_from_es
+
+    retry_until_pass(assert_authors_records_exist_in_es)
+
+    literature_data = faker.record("lit", with_control_number=True)
+    literature_data.update(
+        {
+            "authors": [
+                {
+                    "full_name": "Brian Gross",
+                    "ids": [{"schema": "INSPIRE BAI", "value": "J.M.Maldacena.1"},],
+                    "emails": ["test@test.com"],
+                },
+            ]
+        }
+    )
+    literature_record = LiteratureRecord.create(literature_data)
+    db.session.commit()
+
+    def assert_disambiguation_task():
+        literature_record_from_es = InspireSearch.get_record_data_from_es(
+            literature_record
+        )
+        assert not literature_record_from_es["authors"][0].get("record")
+
+    retry_until_pass(assert_disambiguation_task)
+
+
+def test_disambiguate_authors_doesnt_match_when_author_is_ambiguous(
+    inspire_app, enable_disambiguation
+):
+    author_1 = faker.record("aut", with_control_number=True)
+    author_1.update(
+        {
+            "name": {"value": "Brian Gross"},
+            "ids": [
+                {"schema": "INSPIRE ID", "value": "INSPIRE-00304313"},
+                {"schema": "INSPIRE BAI", "value": "J.M.Maldacena.2"},
+            ],
+            "email_addresses": [{"current": True, "value": "test@test.com"}],
+        }
+    )
+    author_2 = faker.record("aut", with_control_number=True)
+    author_2.update(
+        {
+            "name": {"value": "Brian Gross"},
+            "ids": [
+                {"schema": "INSPIRE ID", "value": "INSPIRE-00300003"},
+                {"schema": "INSPIRE BAI", "value": "J.M.Maldacena.1"},
+            ],
+            "email_addresses": [{"current": True, "value": "test@test.com"}],
+        }
+    )
+
+    author_record_1 = InspireRecord.create(author_1)
+    author_record_2 = InspireRecord.create(author_2)
+    db.session.commit()
+
+    authors = [{"full_name": "Brian Gross", "emails": ["test@test.com"],}]
+
+    literature_data = faker.record("lit", with_control_number=True)
+    literature_data.update({"authors": authors})
+    literature_record = LiteratureRecord.create(literature_data)
+    db.session.commit()
+
+    def assert_disambiguation_task():
+        literature_record_from_es = InspireSearch.get_record_data_from_es(
+            literature_record
+        )
+        assert not literature_record_from_es["authors"][0].get("record")
+
+    retry_until_pass(assert_disambiguation_task)
+
+
+def test_disambiguation_runs_after_lit_record_update(
+    inspire_app, celery_app_with_context, celery_session_worker, enable_disambiguation
+):
+    author_data = faker.record("aut", with_control_number=True)
+    author_data.update(
+        {
+            "name": {"value": "Brian Gross"},
+            "ids": [{"schema": "INSPIRE BAI", "value": "J.M.Maldacena.1"}],
+            "email_addresses": [{"current": True, "value": "test@uw.edu.pl"}],
+        }
+    )
+    author_record = InspireRecord.create(author_data)
+
+    author_data_2 = faker.record("aut", with_control_number=True)
+    author_data_2.update(
+        {
+            "name": {"value": "Test Author"},
+            "email_addresses": [{"current": True, "value": "test123@uw.edu.pl"},],
+        }
+    )
+    author_record_2 = InspireRecord.create(author_data_2)
+    author_data_3 = faker.record("aut", with_control_number=True)
+    author_data_3.update(
+        {
+            "name": {"value": "Another Author"},
+            "email_addresses": [
+                {"current": True, "value": "testxx@uw.edu.pl"},
+                {"current": True, "hidden": True, "value": "testxx@fuw.edu.pl"},
+            ],
+        }
+    )
+    author_record_3 = InspireRecord.create(author_data_3)
+    db.session.commit()
+
+    def assert_authors_records_exist_in_es():
+        author_record_from_es = InspireSearch.get_record_data_from_es(author_record)
+        author_2_from_es = InspireSearch.get_record_data_from_es(author_record_2)
+        author_3_from_es = InspireSearch.get_record_data_from_es(author_record_3)
+        assert author_record_from_es
+        assert author_2_from_es
+        assert author_3_from_es
+
+    retry_until_pass(assert_authors_records_exist_in_es)
+
+    literature_data = faker.record("lit", with_control_number=True)
+    literature_data.update(
+        {
+            "authors": [
+                {
+                    "full_name": "Brian Gross",
+                    "ids": [{"schema": "INSPIRE BAI", "value": "J.M.Maldacena.1"},],
+                    "emails": ["test@uw.edu.pl"],
+                    "uuid": "798d9afe-d3c2-479e-b384-f0aee2573076",
+                },
+            ]
+        }
+    )
+    literature_record = LiteratureRecord.create(literature_data)
+    db.session.commit()
+
+    literature_record["authors"].append(
+        {"full_name": "Test Author", "emails": ["test123@uw.edu.pl"]}
+    )
+    literature_record.update(dict(literature_record))
+    db.session.commit()
+
+    def assert_disambiguation_on_update():
+        literature_record_from_es = InspireSearch.get_record_data_from_es(
+            literature_record
+        )
+        assert literature_record_from_es["authors"][0]["record"]["$ref"]
+        assert literature_record_from_es["authors"][1]["record"]["$ref"]
+
+    retry_until_pass(assert_disambiguation_on_update, retry_interval=2)
