+++ backend/tests/integration-async/migrator/test_migrator_tasks.py
@@ -661,217 +661,3 @@ def test_update_relations_with_modified_experiments(
     ).one()
 
     assert experiment_literature_relation.literature_uuid == record.id
-
-
-def test_disambiguation_runs_after_record_creation(
-    inspire_app, celery_app_with_context, celery_session_worker, enable_disambiguation
-):
-    author_data = faker.record("aut", with_control_number=True)
-    author_data.update(
-        {
-            "name": {"value": "Brian Gross"},
-            "ids": [{"schema": "INSPIRE BAI", "value": "J.M.Maldacena.1"}],
-            "email_addresses": [{"current": True, "value": "test@test.com"}],
-        }
-    )
-    author_record = InspireRecord.create(author_data)
-    db.session.commit()
-
-    def assert_authors_records_exist_in_es():
-        author_record_from_es = InspireSearch.get_record_data_from_es(author_record)
-        assert author_record_from_es
-
-    retry_until_pass(assert_authors_records_exist_in_es)
-
-    literature_data = faker.record("lit", with_control_number=True)
-    literature_data.update(
-        {
-            "authors": [
-                {
-                    "full_name": "Brian Gross",
-                    "ids": [{"schema": "INSPIRE BAI", "value": "J.M.Maldacena.1"},],
-                    "emails": ["test@test.com"],
-                },
-            ]
-        }
-    )
-    literature_record = LiteratureRecord.create(literature_data)
-    db.session.commit()
-
-    def assert_disambiguation_task():
-        literature_record_from_es = InspireSearch.get_record_data_from_es(
-            literature_record
-        )
-        assert (
-            str(author_data["control_number"])
-            in literature_record_from_es["authors"][0]["record"]["$ref"]
-        )
-
-    retry_until_pass(assert_disambiguation_task, retry_interval=2)
-
-
-def test_disambiguate_many_authors_runs_after_record_creation(
-    inspire_app, celery_app_with_context, celery_session_worker, enable_disambiguation
-):
-    author_1 = faker.record("aut", with_control_number=True)
-    author_1.update(
-        {
-            "name": {"value": "Brian Gross"},
-            "ids": [
-                {"schema": "INSPIRE ID", "value": "INSPIRE-00304313"},
-                {"schema": "INSPIRE BAI", "value": "J.M.Maldacena.1"},
-            ],
-            "email_addresses": [{"current": True, "value": "test@test.com"}],
-        }
-    )
-    author_2 = faker.record("aut", with_control_number=True)
-    author_2.update(
-        {
-            "name": {"value": "Donald Matthews"},
-            "ids": [{"schema": "INSPIRE BAI", "value": "H.Khalfoun.1"}],
-            "email_addresses": [
-                {"current": True, "value": "test1@test.pl"},
-                {"current": True, "value": "test1.1@test.pl"},
-            ],
-        }
-    )
-
-    author_record_1 = InspireRecord.create(author_1)
-    author_record_2 = InspireRecord.create(author_2)
-    db.session.commit()
-
-    def assert_authors_records_exist_in_es():
-        author_record_1_from_es = InspireSearch.get_record_data_from_es(author_record_1)
-        author_record_2_from_es = InspireSearch.get_record_data_from_es(author_record_2)
-        assert author_record_1_from_es and author_record_2_from_es
-
-    retry_until_pass(assert_authors_records_exist_in_es)
-
-    literature_data = faker.record("lit", with_control_number=True)
-    literature_data.update(
-        {
-            "authors": [
-                {
-                    "full_name": "Brian Gross",
-                    "ids": [
-                        {"schema": "INSPIRE ID", "value": "INSPIRE-00304313"},
-                        {"schema": "INSPIRE BAI", "value": "J.M.Maldacena.1"},
-                    ],
-                    "emails": ["test@test.com"],
-                },
-                {
-                    "full_name": "Donald Matthews",
-                    "ids": [{"schema": "INSPIRE BAI", "value": "H.Khalfoun.1"}],
-                    "emails": ["test1@test.pl", "test1.1@test.pl"],
-                },
-            ]
-        }
-    )
-    literature_record = LiteratureRecord.create(literature_data)
-    db.session.commit()
-
-    def assert_disambiguation_task():
-        literature_record_from_es = InspireSearch.get_record_data_from_es(
-            literature_record
-        )
-        literature_record_from_es_authors = literature_record_from_es.get("authors")
-        assert (
-            str(author_1["control_number"])
-            in literature_record_from_es_authors[0]["record"]["$ref"]
-        )
-        assert (
-            str(author_2["control_number"])
-            in literature_record_from_es_authors[1]["record"]["$ref"]
-        )
-
-    retry_until_pass(assert_disambiguation_task, retry_interval=2)
-
-
-def test_disambiguate_authors_doesnt_match_when_author_is_ambiguous(
-    inspire_app, enable_disambiguation
-):
-    author_1 = faker.record("aut", with_control_number=True)
-    author_1.update(
-        {
-            "name": {"value": "Brian Gross"},
-            "ids": [
-                {"schema": "INSPIRE ID", "value": "INSPIRE-00304313"},
-                {"schema": "INSPIRE BAI", "value": "J.M.Maldacena.2"},
-            ],
-            "email_addresses": [{"current": True, "value": "test@test.com"}],
-        }
-    )
-    author_2 = faker.record("aut", with_control_number=True)
-    author_2.update(
-        {
-            "name": {"value": "Brian Gross"},
-            "ids": [
-                {"schema": "INSPIRE ID", "value": "INSPIRE-00300003"},
-                {"schema": "INSPIRE BAI", "value": "J.M.Maldacena.1"},
-            ],
-            "email_addresses": [{"current": True, "value": "test@test.com"}],
-        }
-    )
-
-    author_record_1 = InspireRecord.create(author_1)
-    author_record_2 = InspireRecord.create(author_2)
-    db.session.commit()
-
-    authors = [{"full_name": "Brian Gross", "emails": ["test@test.com"],}]
-
-    literature_data = faker.record("lit", with_control_number=True)
-    literature_data.update({"authors": authors})
-    literature_record = LiteratureRecord.create(literature_data)
-    db.session.commit()
-
-    def assert_disambiguation_task():
-        literature_record_from_es = InspireSearch.get_record_data_from_es(
-            literature_record
-        )
-        assert not literature_record_from_es["authors"][0].get("record")
-
-    retry_until_pass(assert_disambiguation_task)
-
-
-def test_disambiguation_doesnt_run_with_future_flag_disabling_it(
-    inspire_app, celery_app_with_context, celery_session_worker
-):
-    author_data = faker.record("aut", with_control_number=True)
-    author_data.update(
-        {
-            "name": {"value": "Brian Gross"},
-            "ids": [{"schema": "INSPIRE BAI", "value": "J.M.Maldacena.1"},],
-            "email_addresses": [{"current": True, "value": "test@test.com"}],
-        }
-    )
-    author_record = InspireRecord.create(author_data)
-    db.session.commit()
-
-    def assert_authors_records_exist_in_es():
-        author_record_from_es = InspireSearch.get_record_data_from_es(author_record)
-        assert author_record_from_es
-
-    retry_until_pass(assert_authors_records_exist_in_es)
-
-    literature_data = faker.record("lit", with_control_number=True)
-    literature_data.update(
-        {
-            "authors": [
-                {
-                    "full_name": "Brian Gross",
-                    "ids": [{"schema": "INSPIRE BAI", "value": "J.M.Maldacena.1"},],
-                    "emails": ["test@test.com"],
-                },
-            ]
-        }
-    )
-    literature_record = LiteratureRecord.create(literature_data)
-    db.session.commit()
-
-    def assert_disambiguation_task():
-        literature_record_from_es = InspireSearch.get_record_data_from_es(
-            literature_record
-        )
-        assert not literature_record_from_es["authors"][0].get("record")
-
-    retry_until_pass(assert_disambiguation_task)
